// src/app/api/model/cashflowPrediction.ts
/**
 * SmartLedger API
 *
 * Contact: support@smartledger.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Organization } from './organization';


export interface CashflowPrediction { 
    id?: string;
    orgId?: string | null;
    month?: string;
    projectedInflow?: number | null;
    projectedOutflow?: number | null;
    confidence?: number | null;
    generatedAt?: string | null;
    org?: Organization;
}